Add more features to the GUI.
- new dilog for setting algorithm variables
- checking input values
- resize canvas with parent window
- slightly changed colors

+++ src/mcoufal/devel/aco/core/Ant.java
@@ -41,7 +41,7 @@ public class Ant {// TODO: check if arraylist is OK with my cause(order of items
 	// current path length
 	private double pathLength = 0;
 	// pheromone quantity given by each ant per cycle
-	private int quantityOfPheromone = 100;
+	private double quantityOfPheromone = 100;
 
 	public Ant(AntColonyAlgorithm algorithm, ArrayList<CrossroadPoint> allCrossroads) {
 		this.numberOfCrossroads = allCrossroads.size();
