+++ src/mcoufal/devel/aco/core/AntColonyAlgorithm.java
@@ -3,8 +3,6 @@ package mcoufal.devel.aco.core;
 import java.util.ArrayList;
 
 public interface AntColonyAlgorithm {
-	// TODO comments
-	// TODO define all needed functions
 	public void step();
 	public void run();
 	public ArrayList<CrossroadPoint> getBestPath();
@@ -14,8 +12,8 @@ public interface AntColonyAlgorithm {
 	public double[][] getTau();
 	public double[][] getDeltaTau();
 	public double[][] getEta();
-	public int getAlpha();
-	public int getBeta();
-	public int getQuantityOfPheromone();
+	public double getAlpha();
+	public double getBeta();
+	public double getQuantityOfPheromone();
 	public double getRo();
 }
